// src/types/openapi/models/ObjectEventHistoryDto.ts
// tslint:disable
/**
 * Core API
 * REST API for CZERTAINLY Core
 *
 * The version of the OpenAPI document: 2.17.1-SNAPSHOT
 * Contact: info@otilm.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { NameAndUuidDto, ResourceEvent, ResourceObjectDto, TriggerHistoryRecordDto } from './';

/**
 * @export
 * @interface ObjectEventHistoryDto
 */
export interface ObjectEventHistoryDto {
    /**
     * Resource event that was triggered.
     * @type {ResourceEvent}
     * @memberof ObjectEventHistoryDto
     */
    event: ResourceEvent;
    /**
     * Trigger that evaluated the event.
     * @type {NameAndUuidDto}
     * @memberof ObjectEventHistoryDto
     */
    trigger: NameAndUuidDto;
    /**
     * Origin that caused the event to be triggered.
     * @type {ResourceObjectDto}
     * @memberof ObjectEventHistoryDto
     */
    origin?: ResourceObjectDto;
    /**
     * All conditions in the trigger have been matched.
     * @type {boolean}
     * @memberof ObjectEventHistoryDto
     */
    conditionsMatched: boolean;
    /**
     * All actions in the trigger have been performed.
     * @type {boolean}
     * @memberof ObjectEventHistoryDto
     */
    actionsPerformed: boolean;
    /**
     * Time at which the event was triggered.
     * @type {string}
     * @memberof ObjectEventHistoryDto
     */
    triggeredAt: string;
    /**
     * Additional message.
     * @type {string}
     * @memberof ObjectEventHistoryDto
     */
    message?: string;
    /**
     * If trigger was sending notifications, confirmation whether notification has been sent successfully or not.
     * @type {boolean}
     * @memberof ObjectEventHistoryDto
     */
    notificationsSent?: boolean;
    /**
     * List of records for each condition that has not been matched and each action that has not been performed.
     * @type {Array<TriggerHistoryRecordDto>}
     * @memberof ObjectEventHistoryDto
     */
    records: Array<TriggerHistoryRecordDto>;
}